Extract website module from main
