+++ src/main.rs
@@ -2,6 +2,7 @@
 
 mod api;
 mod store;
+mod website;
 
 #[macro_use]
 extern crate rocket;
@@ -12,117 +13,13 @@ extern crate diesel;
 #[macro_use]
 extern crate diesel_migrations;
 
-use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 
-use chrono::prelude::*;
-use maud::{html, Markup, DOCTYPE};
 use rocket::fairing::AdHoc;
 use rocket::response::NamedFile;
 use rocket::State;
 
-use store::{
-    Actions, Event, Id, Location, OccurrenceFilter, OccurrenceWithEvent, OccurrenceWithLocation,
-    Store,
-};
-
-#[get("/")]
-fn index(store: Store) -> Markup {
-    html! {
-        ( DOCTYPE )
-        html lang="de" {
-            head {
-                meta name="viewport" content="width=device-width, initial-scale=1";
-
-                link href="static/main.css" rel="stylesheet";
-            }
-            body {
-                header {
-                    h1 { "Lindy Hop Aachen" }
-                }
-                main {
-                    ol.schedule {
-                        @let locations: HashMap<Id<Location>, Location> = store.all();
-                        @for occurrences_for_date in store.occurrences_by_date(&OccurrenceFilter::upcoming()) {
-                            li { ( render_entry(&occurrences_for_date, &locations) ) }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn render_entry(
-    (date, entries): &(NaiveDate, Vec<OccurrenceWithEvent>),
-    locations: &HashMap<Id<Location>, Location>,
-) -> Markup {
-    html! {
-        div.date { ( format_date(date) ) }
-        ol.events {
-            @for occurrence_entry in entries {
-                li.event { ( render_occurrence(occurrence_entry, locations) ) }
-            }
-        }
-    }
-}
-
-fn format_date(date: &NaiveDate) -> String {
-    use chrono::Weekday::*;
-
-    let day = match date.weekday() {
-        Mon => "Mo",
-        Tue => "Di",
-        Wed => "Mi",
-        Thu => "Do",
-        Fri => "Fr",
-        Sat => "Sa",
-        Sun => "So",
-    };
-    let format = format!("{}, %d.%m.", day);
-
-    date.format(&format).to_string()
-}
-
-fn render_occurrence(
-    entry: &OccurrenceWithEvent,
-    locations: &HashMap<Id<Location>, Location>,
-) -> Markup {
-    html! {
-        @let entry_html =  html_from_occurrence(&entry.occurrence, &entry.event, locations);
-        div.quick-info { ( entry_html.quick_info ) }
-        h2.title { ( entry_html.title ) }
-        div.content {
-            div.description {
-                div.teaser { ( entry_html.teaser ) }
-            }
-        }
-    }
-}
-
-struct OccurrenceHtml {
-    title: Markup,
-    quick_info: Markup,
-    teaser: Markup,
-}
-
-fn html_from_occurrence(
-    occurrence: &OccurrenceWithLocation,
-    event: &Event,
-    locations: &HashMap<Id<Location>, Location>,
-) -> OccurrenceHtml {
-    let maybe_location = locations.get(&occurrence.location_id);
-    let location_name = match maybe_location {
-        Some(location) => &location.name,
-        None => "Steht noch nicht fest.",
-    };
-
-    OccurrenceHtml {
-        title: html! { ( event.title ) },
-        quick_info: html! { ( format!("{} - {}", occurrence.occurrence.start.format("%H:%M"), location_name) ) },
-        teaser: html! { ( event.teaser ) },
-    }
-}
+use store::Store;
 
 #[get("/admin")]
 fn admin_route() -> Option<NamedFile> {
@@ -147,7 +44,6 @@ struct AssetsDir(PathBuf);
 #[get("/static/<file..>")]
 fn static_file(file: PathBuf, assets_dir: State<AssetsDir>) -> Option<NamedFile> {
     let path = assets_dir.0.join(file);
-    print!("{:?}", path);
     NamedFile::open(path).ok()
 }
 
@@ -167,9 +63,8 @@ fn main() {
                 Err(rocket)
             }
         }))
-        .mount(
-            "/",
-            routes![static_file, index, admin_route, admin_subroute],
-        );
-    api::mount(rocket, "/api").launch();
+        .mount("/", routes![static_file, admin_route, admin_subroute]);
+    let rocket = website::mount(rocket, "/");
+    let rocket = api::mount(rocket, "/api");
+    rocket.launch();
 }
